backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.0.0,<2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
postgrest==0.13.0
httpx==0.24.1
supabase==2.3.0

